refactor `getStateReplay`
